Updating domain model image to account for renewed Puppet-classifier discussions.
The puppet node classifier is of limited scope currently -- puppet handles much
of the dependency concerns in manifest files, requires only a flat "class" list,
mapped to nodes.  Connection to customer modeling is nebulous currently.

The domain has been updated to highlight which portions of the current model are
germain to larry, to the puppet classifier, and to both.