Relax rake requirement to >= 0.8.3
A dependency on rake >= 0.8.4 was added to the Dashboard to fix issue #2897,
caused by a missing reenable method on Rake::Task. Rake's release notes [1]
indicate that this method was added in 0.8.2. The Rails frozen in vendor/
require rake >= 0.8.3. Since EPEL currently ships Rake 0.8.3, it would be nice
to relax our dependency to that version.

[1] http://rubyforge.org/forum/forum.php?forum_id=27141