[#3298] Increase the size of the reports.report column from 64k to 16mb
This should prevent MySQL from truncating reports, although database
size may balloon quickly.

The migration in this patch works on MySQL 5, but probably not MySQL 4,
but that should be okay since in the README it says we require MySQL 5.
Strangely,  putting mediumtext in the field type in the migration makes a
longtext field, but putting longtext in the migration didn't expand the
field size at all.

16mb seems to be limit of the text field size without modifying
max_allowed_packet for MySQL.  Past 16mb it doesn't truncate data
anymore, but errors since it runs into the max_allowed_packet limit.