[#4187] Group graphs containing the same group twice aren't considered cyclic
Group graphs were being marked as cyclic when they simply contained the same
group twice from different paths. This patch fixes that, and also moves the
cycle detection to be validation on the NodeGroupEdge class, so that cycles
are detected and prevented when they are created. Previously, this detection
was being done incidentally when the node page was viewed, and was throwing
a cycle error which caused the page to be unrenderable.