[#4503] Make the class name regex more like puppet's
Class name regex can be found in lib/puppet/parser/lexer.rb:155
%r{([a-z][–\w])?(::[a-z][–\w])+} => :CLASSNAME, # Require ‘::’ in the %class name, else we’d compete with NAME

This makes dashboard use a similar regex.  It's different in that in
dashboard it still doesn't allow you to create a class name with :: in
front of the class name.  There's currently a bug or at least an
inconsistent behavior in puppet (#4523) that deals with the preceding ::

Another change from what dashboard used to have is that capital letters
aren't allowed.  Jesse mentioned that there's rumbling that puppet may
change to allow caps, so when that changes, we'll want to change
dashboard.  It's unfortunate that it's hard to keep dashboard name rules
in sync with puppet's, but I can't think of a good way to do this
automatically.