Remove unnecessary report loading from pages controller
The @reports variable is not used, and (in at least some versions of Rails)
the :order => 'time ASC' used to generate it causes an expensive database
query which sorts by time first in ascending, then in descending order (due to
the ordering on Report's default_scope). Remove this query entirely to improve
performance.