[#4008] Fixed group parameter inheritance
Group parameters were failing to inherit properly because premature
caching was disrupting the recursive compiled_parameters method. This
patch entails a rewrite of compiled_parameters, which fixes several
other issues with incorrect parameter inheritance. It also fixes
the problem of a node with conflicting parameters failing when a
user tries to view it in HTML. It will now give a warning message,
but will still fail if accessed via the external node tool.