[#4403] Run failure chart now shows the correct reports
Reports are stored in UTC, and therefore the run failure chart was
grouping the runs for days starting at UTC midnight, which doesn't
make sense. Now it uses localtime days, but also now only supports
grouping reports for days, rather than an arbitrary interval. That
capability should probably be put back in eventually, and made to
actually work, when we have a need for it.