[#4503] Fix NodeClass#name validation regexp to be more like Puppet's.
In Puppet, "lib/puppet/parser/lexer.rb" at line 155 says the regexp is:
  %r{([a-z][-\w]*)?(::[a-z][-\w]*)+}

Made the validation regexp closer to this, but prevented leading colons
because this is illogical in Dashboard since that would be the same as
having a class without the leading colons.

Expanded specs to forbid blanks and trailing colons, but allow
single-letter classes.