[#4578] Node classifier properly returns an empty node when node not found
Previously, a node not found would trigger a 404 error, but now it will
return YAML representing an empty node (an empty list of classes, no
parameters and no name). This is to allow Puppet to use its own node
classification in the event a node is not classified by Dashboard.