[#3470] Switch class, and group tokenizer/autocomplete to use new JavaScript library.
The old library had issues rendering properly in Firefox, and would only
add classes/groups if clicked on.  Switching to this new library gets rid
of these issues.

While we're switching to the new autocomplete library, switch the Node,
and NodeGroup models to accept class, and group ids from the edit form,
instead of names.

This change also addresses issues #3427, and #4635.