[#4625] Moved pagination from views into controllers
This fixes the problem of every paged index first retrieving all
records, ignoring them, and then retrieving the page it actually wants.
In the case of reports, this was taking a very long time (on the order
of seconds), and is very undesirable. All controllers will still return
the full dataset when requested in YAML or JSON form.