[#3470] Add find_from_form_* class methods to NodeClass & NodeGroup
These find_from_form_names, and find_from_form_ids helpers consolidate the
logic used to parse the values we get from the tokenizer/autocomplete
JavaScript.

This could be consolidated further, but this at least cuts the duplication
in half from what it would be otherwise.

This also addresses bug #4635