Standardize whitespace used for indentation, and wrap longer code lines
Text for lists should start at columns that are multiples of 4.

    1.  Foo
    VS
    1. Foo

This makes it easier to properly indent things within the lists (code
blocks, and sub-lists), since they will also start at a multiple of 4,
instead of having to make sure that they start with the _text_ from the
line above.  Both conditions of "indents are 4 spaces", and "sub-indents
start from the left edge of text in lists" will be satisfied naturally.