[#4801] Fix exception in node's report listing.
The NodesController#report action was failing because it tried to
render a raw array as if it was a paginated array.

This error was introduced in 8124c51b as part of #4625 "Reports index
should only fetch the data it needs" which had an insufficient
refactoring of the pagination.

The existing tests didn't fail because the value was never rendered.
The controller specs set the value, but didn't render it; while the
view specs set their own paginated value and rendered it.

Solution:
* Added `integrate_views` to NodesController to force rendering of
  templates, which caused them to start failing as expected.
* Moved #paginate_scope from ApplicationHelper to PaginateScopeHelper
  and included it in ApplicationController so that all controllers
  could add pagination to their result sets.
* Added pagination to the NodesController#report action for HTML.
* Refactored controllers and mixins to use #paginate_scope, rather
  than reinventing it themselves.
* Eliminated unnecessary #paginate_scope :per_page option.
* Eliminated unused #items_per_page helper.

Signed-off-by: Igal Koshevoy <igal@pragmaticraft.com>