(#4811) Fix raw URLs, which break when web app is mounted in a subpath.
By default, the app runs without a subpath, so that the homepage is at "/"
and nodes are at "/nodes". However, some users need to mount the app under
a subpath so that, for example, the homepage is at "/dashboard/" and the
nodes are at "/dashboard/nodes".

For this to work, the application must use relative or generated URLs
whenever it's referring to itself.