(#4825) Fix raw "favicon.ico" URL, which breaks if app is mounted in subpath.
By default, the app runs without a subpath, so that the
"/images/favicon.ico" path is correct. However, some users need to mount
the app under a subpath so that, for example, this path becomes
"/dashboard/images/favicon.ico".

The solution is to use Rails routing-aware helpers to produce paths.