Fixed RPM spec file to use more compatible way of detecting user.
The spec used `root-%(%{__id_u} -n)`, which fails on Ubuntu 9.04 and on
systems using an older version of `rpmbuild`, and tries to include the
name of the user twice -- once hardcoded and again using interpolation.

Changed this to use `%(id -un)` which sets the username once and is
compatible with older versions of `rpmbuild`.