(#2933) Add table of facts to node pages
Facts are asynchronously queried from an https location, specified
with the configuration options "inventory_server" and "inventory_port".
Added a route /nodes/NODENAME/facts which will render the table of
facts, and is used for the NodesController#facts action, which does
the asynchronous query to the inventory service and renders the table
of facts.