(#4401) Modified status query to use the Rails time zone
The status query (which we use to generate charts) was using the
Time.now and Time.parse methods, which use the system time zone.
Changed them to use Time.zone.now and Time.zone.parse, which use the
Rails time zone (which we allow the user to configure in
settings.yml).

Paired-with: Nick Lewis <nick@puppetlabs.com>