(#5459) Convert from yaml in the report column to a schema
Storing most of the report data in the database as serialized yaml has
been a performance bottleneck for some time.  Deserializing that data
every time something inside needed to be accessed was costly and
prohibited some features we have planned for the future including the
ability to search on some of that data.

This commit series will convert that data into a schema.  The conversion
process can be very time consuming for databases that have a large
number of reports, so the migration to do this will first backup the
reports table and create a new reports table along with the new tables
(logs, resource_statues, metrics, events) without any data.  Then the
old reports can be migrated into the new tables with a rake task as a
background process.  This rake task will be created as a separate task
for this iteration.

Paired-with: Matt Robinson and Paul Berry