(#5535) Rake task to migrate old database yaml to the new reports schema
As each report from the old_reports table is saved to the new reports
schema, it is deleted from the old_reports table.  This means you can do
this migration in chunks and resume when it's convenient.  Recent
reports are migrated first under the assumption that more recent data is
more relevant.  This move is done transactionally so that interrupting
the migration won't leave the database in an inconsistent state.

Reviewed-by: Nick Lewis