Fix #5573.  Diff now handles missing resources/properties.
Previously, the report diffing algorithm would raise an exception or
ignore differences when a resource was present in one report but
absent in the other, or a property was present in one report but
absent in the other.  Now the diffing algorithm generates "nil"s to
represent "not present".  This renders in the view as empty table
cells.

Paired-with: Jesse Wolfe