(#5725) Remove user related code
There’s a lot of code in Dashboard related to users, and none of it is being
used. It looks like user code was added despite never really working or being
required. Users are surely a goal in some form for dashboard – someday. But for
now it’s just cruft. Also, if we want to upgrade to Rails 3 before we add
users, this crufty code makes it harder.