(#5866) Get a consolidated report for baseline diffs in a group
You can already get a diff of a node's most recent inspect report
against it's baseline, but it's a little tedious to get a lot of nodes
diffs back in bulk.  This feature makes it easier in that you can put
nodes into groups and all with one button click get back the diff of the
latest inspect report for a node against the baseline report for that note.

Reviewed-by: Nick Lewis