Revert "Maint: Removed show_expand_all_link variable"
This reverts commit 353998f22a2c4981fb50b1b564d025b0bf9fc0a0.

Removing the show_expand_all_link from the _diff partial causes the
expand links on the group diff pages to go crazy.  It may be that we
want a way to have an expand all link on a per node basis on these
pages, but this will require smarter naming of our divs so that the
expand all links on each node don't interact.  Until that time, I'm
reverting this patch.

Reviewed-by: Paul Berry