maint: add a view test to motivate reverting diff expand_all
I'm not sure about the value of tests like this, but wanted to try
writing one with the goal of preventing regressions like the one that
accidentally happened with the expand_all link.  It may turn out that
view tests don't add much value or catch these sorts of regressions, or
that we need a different level of view testing with something like
Selenium or webrat, but we'll have to experiment to see what works.

Reviewed-by: Jesse Wolfe