maint: Fix a case where the alphabetically first baselines may not appear
The /reports/baselines query was retrieving n reports from the database, using
the default ordering by time. This wasn't caught by the test because the test
happened to be creating reports in such an order that the three it expected
were the three retrieved. Fixed it to properly sort by name to select the n
alphabetically first reports.

Reviewed-By: Paul Berry