(#6069) Fixed unique ids in the report group diff view.
When diffing all the nodes in a group against baselines, the "expand"
links were failing to work properly because every time the _diff
partial was executed, it re-used id's that were supposed to be unique
on the page.  Changed the mechanism for generating unique ids to a
method generate_unique_id() in the application helper, so that the ids
will always be unique on the page.

Paired-with: Nick Lewis <nick@puppetlabs.com>