Maint: removed db/schema.rb
The file db/schema.rb tracks the current state of the database.  In a
modular dashboard, this state isn't a meaningful thing to track in
source control, because it depends which modules are currently
installed.  Also, the information in the file is redundant since it
can always be regenerated using "rake db:migrate".  Accordingly, we
are remoing it from source control.

After pulling down this change, you will have to run "rake db:migrate
RAILS_ENV=test" once in order to build a local copy of db/schema.rb so
you can run spec tests.  This should only be necessary once.

Paired-with: Jesse Wolfe <jesse@puppetlabs.com>