Maint: Moved logic for identifying inspect reports into a callback.
The logic for marking inspect reports with an appropriate icon is now
in a callback method that plugs into :report :status_icon, rather than
being hardcoded in the _report_status_icon partial.  This paves the
way for additional future modularity in choosing report icons.

Paired-with: Nick Lewis <nick@puppetlabs.com>