(#6835) Handle malformed facts from Puppet 2.6.7 & storedconfigs
When querying Puppet 2.6.7 and storedconfigs are enabled along with the
inventory service, the timestamp information comes back as a "special"
fact in the values section rather than as metadata in the top level.

Now we handle pulling out this "special" fact, along with being able to
cope with no timestamp coming back in the metadata section, or as this
"special" fact.

Paired-with: Jesse Wolfe <jesse@puppetlabs.com>