(#7209) Require naming convention for plugin migration files
We require plugin migrations to be named in the form
YYYYMMDDHHMMSS_plugin_#{plugin_name}_*.rb to avoid any class
name and filename collisions between plugins and core Puppet
Dashboard or between plugins and other plugins.

We can't simply rename the migration as we copy it into
place, since this would require modifying the file contents
to match the new name, and down that road lies madness.

Requiring module authors to adhere to this naming convention
for their migrations is fairly low cost given the previously
mentioned constraints.

Reviewed-by: Jesse Wolfe <jesse@puppetlabs.com>