(#5126) Support HTTP basic auth in bin/external_node
Although URLs with usernames and passwords were valid in the previous version of the external_node
script, they weren't being passed in the HTTP request. This commit adds a username and password
to the request if a user is present in the URL.

The "if uri.user" isn't strictly necessary, since requests with blank usernames/passwords
would still succeed if basic auth isn't enabled in Dashboard, but it adds clarity.

Reviewed-by: jacob@puppetlabs.com