(#6991) Massage Node named scopes.
In support of the node status summary change, it is valuable for us to have
a few named scopes that we can easily compose.  To that end, we've refactored
the pending scope to take a boolean predicate, and introduced two new scopes
('current' and 'successful').  While these scopes bear only passing resemblance
to the existing concepts with those names, we are moving towards these
semantics overall.

Paired-With: Nick Lewis