maint: Have different production and development databases in the example
The database.yml.example has both production and development databases
set to be 'dashboard'.  If you just copy this and run `rake db:create:all`
you'll get an error, and if you very mistakenly use the same database
for development and production very bad things will happen.

This just changes them to be different in the example file so that it's
easier to get started with sane defaults.

Reviewed-by: Josh Cooper <josh@puppetlabs.com>