maint: Add resource status scopes for pending/failed
These are similar to the corresponding scopes on node, in that they accept
predicates for failed=true/false and pending=true/false. We do this with a
scope for failed instead of using a finder (find_all_by_failed) so we can chain
it with other scopes for lazy evaluation. These scopes are necessary for CSV
export support.

Paired-With: Ben Hengst