(#7560) Clean up the Dashboard index for 0 nodes.
The case where Dashboard contains no nodes is an unfortunately common case;
every new Dashboard installation comes in such a state.  Prior to this change,
the user experience was that they could not actually view Dashboard in their
browser before at least one node had reported in.

This change replaces the index page with a more informative message when
Dashboard is not aware of any nodes.

Reviewed-By: Jacob Helwig