(#7007) Add support for CSV output
This adds a CsvSerializable module which is mixed into Array,
ActiveRecord::NamedScope::Scope, and adds functionality to
ActiveRecord::Base implementing the to_csv method.

On Array, and ActiveRecord::NamedScope::Scope this method will call
"to_csv_header" on the class of the first element of the array to get the
header, and then "to_csv" on each element. If a block is given, each line will
be yielded to the block, otherwise, the lines will be concatenated and
returned.

With ActiveRecord::Base:

 * #to_csv will export the values provided by Model#to_csv_array.

 * Model#to_csv_array defaults to the values of each property specified by
   Model::to_csv_properties.

 * Model::to_csv_properties defaults to the column names of the model table.

 * Model::to_csv_header defaults to using Model::to_csv_properties as the
   header.

Paired-With: Daniel Pittman
Paired-With: Jacob Helwig <jacob@puppetlabs.com>