(#7007) Use a custom property list for ResourceStatus.to_csv
When exporting ResourceStatus as CSV, we don't want to include id, report_id,
and tags, so we override the default to_csv_properties.

We exclude id and report_id because they are completely internal fields to
Dashboard. Tags are excluded because, as an array, they don't export well to
CSV.

Paired-with: Jacob Helwig <jacob@puppetlabs.com>