(#7007) Stream node CSV download
Before, when attempting to download the CSV export of nodes and their resource
statuses, the download wouldn't start until after all of the data had been
collected, and converted to CSV.

By passing render a proc, and using output.write, we can send each CSV line to
the client as it is generated.

The main reason for doing this is to give the user immediate feedback that we
recognize they clicked on the download link, instead of stalling until all of
the data has been collected.  This also should reduce the server-side memory
footprint of the CSV, since we don't need to generate the entire CSV string in
memory before sending it.

Paired-with: Jacob Helwig <jacob@puppetlabs.com>