(#7630) Denormalize ResourceStatus status and count calculations
Initially the plan for this ticket was to do the denormalization columns
in the report transformer, and then update puppet to output the new
data.  However, after some discussion with Randall about where we'd like
to head with the report format that Puppet outputs, we'd like to move
away from including anything but the bare minimum on the reports that
puppet generates.

So instead we're calculating the status counts after the report
transformer does its work.  The downside to this is that there's yet
another way to 'transform' report data, and that Dashboard is diverging
from what Puppet sends out for reports.  We're willing to make this
trade-off right now while we think about how we want the ideal Puppet
report format to look.

We're also adding the pending status to report statuses.  Currently
report statuses can only be unchanged, changed, and failed.

The db migration only adds the new count data for the most recent report
for each node only, since migrating every report could be really slow
for large databases unless done in complicated sql, and the counts being
calculated are only planned to be displayed for the most recent reports.
Should anyone want all their reports updated, there's a `rake
reports:update_report_data` task added that's idempotent, so I can be
run multiple times without messing up the data.

Paired-with: Max Martin