(#7674) Add pending to the run status stacked bar graph
The bar chart was using the same color for changed reports as the count
summary was using for pending.  Now the colors are consistent with each
other, and all the statuses are represented.

We may need to pick friendlier colors, both for color blind and general
niceness to the eye, but that can be tweaked easily later.

Paired-with: Max Martin <max@puppetlabs.com>