maint: Fix migration to check for nil value
Our latest migration to add resource status information to the metrics
tables was calling the munge method on reports before checking to see
whether the node actually had a report. This commit adds a sanity check
to make sure munge won't be called on a nil object.

Paired-with:Matt Robinson <matt@puppetlabs.com>