(#7913) Upgrading from Rails 2.3.4 to 2.3.12
Upgrading the vendored version of Rails helps with both performance and
security vulnerabilities.  This also gets us closer to a Rail 3 upgrade
being smoother since we'll fix deprecation warnings.

The specific performance problem that motivated this change is that
doing nested attribute record creation is more than twice as fast.  Not
sure exactly what ActiveRecord was doing so poorly before, but we were
playing with solutions that involved a lot less maintainable code since
we had to do a lot of work without active record to trick it into
performing well.

The version of rack that we vendored also needed to be updated from
1.0.1 to 1.1.2 since vendor/rails/actionpack/lib/action_controller.rb
explicitly requires "~>1.1.0".  If the a different version is loaded
first you get errors along the lines of:

    can't activate rack (~> 1.1.0, runtime) for [], already activated rack-1.0.1 for []

Reviewed-by: Pieter van de Bruggen <pieter@puppetlabs.com>