(#7689) Rake task to support parallel report POSTing.
It is sometimes convenient to be able to throw a whole pile of reports into
the import process through the regular upload process.  This is more useful
for testing than uploading old reports, given the parallel import process.

Usage is:

    rake reports:repost REPORT_DIR=/path/to/spool CONCURRENCY=16 \
         # REPORT_URL=http://dashboard:3000/reports/upload

Concurrently controls the maximum number of concurrent threads POSTing at one
time, and is used to implement the client-side concurrency control.