maint: Remove db/seeds.rb so that db:reset task works
The data that was in the seed file was not something that was necessary
for a fresh install of the database.  Besides that, it was using a
generate method that was only available in testing for generating test
data with ObjectDaddy or FactoryGirl.

Paired-with: Pieter van de Bruggen <pieter@puppetlabs.com>