(#7398) Use DelayedJob for background processing.
Delayed_Job is a gem that implements robust background processing for Ruby
tasks.  Now we have vendored it, we can start to use it to process report
import in the background.

This has two advantages: one, we can import in parallel, since we can have
more than one worker in the background working on the YAML transformation and
database updating.

Two, we now return to the report submitter in short order: all we have to do
is spool the YAML to disk, and register the background job, and we are assured
that we will eventually get that into the database.

This eliminates a lot of the master => dashboard submission pipeline delay, so
that the master gets back to serving clients immediately, rather than having
to wait until the report is entirely ingested.

Reviewed-By: Daniel Pittman <daniel@puppetlabs.com>
Reviewed-By: Matt Robinson <matt@puppetlabs.com>