(#7398) Configurable DelayedJob worker count.
We added a setting to configure the number of DelayedJob workers, defaulting
to two, and use that when we fire up the monitor.  Using two workers is a
reasonable default for finding problems caused by actual concurrency in
delayed work, without overloading too much systems with only a single core.

Also, update documentation to recommend that this is tuned to one worker per
core, which is the optimal saturation for long term use.  (At least, in the
current CPU-dominated use we make of them.  If we get I/O bound later we might
need to increase that.)

Reviewed-By: Matt Robinson <matt@puppetlabs.com>