(#7389) Cheaper unique filenames for spooled reports.
Rather than using a content-checksum to name the files, we can just use the
combination of a PID and serial number to give a very high chance of being
unique in a single process.

(This will typically fail only if we have multiple concurrent threads and
 lose a race on the lock-free counter, or have wrapped pids and had spooled
 reports from a previous instance with the same pid.)

We ensure uniqueness using O_CREAT|O_EXCL, so we have the assurance from the
OS that the race will never result in overwritten files, just an exception and
retry.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>