(#7389) Don't auto-start DelayedJob workers.
The code used to automatically fire of DelayedJob workers work trigger every
time Rails was loaded, including when a DelayedJob worker initialized.  This
had the obvious fork-bomb effect...

On inspection it turns out that the different process models of Webrick,
Passenger, and Unicorn, make doing this in a robust and effective way
difficult.

Instead, we back out that code and document the need to run a worker process,
along with the ways to have that running.  This puts responsibility on the
user to get a worker in place.

Paired-With: Matt Robinson <matt@puppetlabs.com>