(#7976) Node filter links in sidebar work in all cases
Previously we were relying on some bad markup and tricky JavaScript to show
correct hover states and make the entire table row work as a link.  This markup
is cleaner and more semantic, and the JavaScript is simpler.  We lose a small
amount of niceness in the presentation, and the CSS is a little hairier, but
that's a worthy tradeoff.

Reviewed-by: Pieter van de Bruggen <pieter@puppetlabs.com>