(#7967) Infrastructure for displaying background failures.
Now that we perform various tasks, such as report imports, in the background
we don't have any immediate way to give users feedback on the outcome of their
task.

This adds infrastructure for collecting and displaying those problems to the
user as part of the normal operation of the dashboard.

The UI is partially completed; all the data is present, but the design is
awful.  This merges in sufficient functionality for our UX team to work on.

Paired-With: Josh Cooper <josh@puppetlabs.com>
Paired-With: Matt Robinson <matt@puppetlabs.com>