(#8022) Provide a rake task to symlink public assets from plugins into `public`.
This allows Dashboard plugins to include CSS and Javascript files that can
easily become a part of the more integrated site.  To take advantage of this,
plugin developers should create a `public` directory in the top level of
their plugin, and populate it with static assets. The rake task will
recursively symlink anything in the `public` directory of a plugin into
the top-level Dashboard `public` directory, creating any missing
directories in the tree.

Paired-With: Max Martin <max@puppetlabs.com>
Reviewed-by: Nick Lewis <nick@puppetlabs.com>