(#8146) Change default DASHBOARD_URL in external_node to localhost
In the most common scenario, in which Dashboard and puppet master are installed
on the same system, Dashboard was unable to be used as an external node
classifier without editing the script or changing the system's name resolution.
This commit allows the most common case to work silently, by changing the
default DASHBOARD_URL from http://dashboard:3000 to http://localhost:3000.