(Maint) More generalized tabbed interface
It was bugging me that the tabs didn't behave correctly.  That is, the mouse
cursor changed to a pointer on areas of the tab that were not clickable.  I
solved this by displaying the anchor tag contents of each tab as a block
element, and removing padding from the tab itself.

I also removed specific ids from the CSS, leaving only a "tabbed" class.