(#8121) Properly generate CSS from SASS in production.
According to this link (http://blog.citrusbyte.com/2008/03/20/sass-production-woes-in-rails/),
Sass has known issues with CSS generation in Rails 2.3.  This is
remedied easily enough by providing an initializer that forces the
plugin to "reload" a little later in the process, but the exact
mechanism isn't completely understood.  This is an interaction
between the Haml and Sass gems, since Haml vendors Sass.  When we
upgrade Haml to 3.2.0, we should be able to remove this
initializer.

This change may negatively impact the first page load on a new
server, but the change is fairly minimal.

Reviewed-By: Matt Robbinson