(#8266) Back-end logic for splitting read and unread DJ failures.
Rather than displaying a single, giant list of delayed_job failures on this
page, we want to split it out into two lists: one unread, which is important,
and one read, which isn't so much.

This does the basic controller logic to support displaying those two pages,
fetching the right content from the database, paginating as appropriate, and
generally getting data to views in the right order.