(#7967) Improved user-facing design for delayed job warnings
Now that Dashboard depends on background tasks ("delayed jobs") to import
reports, we're showing users the status of these tasks.  We improved the sidebar
notification section with status icons.  We improved the page to which they link
by (a) making a tabbed view, and (b) making it a little shinier.

Reviewed-by: Pieter van de Bruggen <pieter@puppetlabs.com>