(#8199) Move 'failed' resources to the top when viewing report events
Because failed resources are the resources a user likely cares about most, we
move those to the top to make them more immediately visible.

This commit also moves the logic for grouping and sorting resource_statuses
from the view to the callback. Because the callback functions as an ad-hoc
controller for the partial it renders, it is a better place for this logic than
the view itself.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>