(#8262) Create callbacks for each section of node_classification partial
We no longer consider groups to be strictly part of the node classification
partial, so it doesn't belong in there anymore. However, we still want groups
to appear before classes, so the best way to accomplish this is by splitting
parameters, groups, and classes into their own separate callbacks, so that we
can make parameters and classes conditional, but always include groups. This
template was only rendering those other partials, so there is no real loss
here.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>