(#8262) Show node groups even when node classification is disabled
Because node groups are meaningful beyond the scope of node classification, we
want to show them in all cases, node just when node classification is enabled.
Parameters and classes are still not shown when node classification is
disabled, and nodes can't be accessed via YAML.

This also entails allowing both nodes and node groups to be edited, to allow
nodes to be put into groups. This was previously disabled because the
use_external_node_classification feature initially doubled as a read-only mode.
However, given the existence of the enable_read_only_mode setting, it is now
acceptable to relax this restriction, and allow groups and nodes to be edited
even with node classification disabled.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>