(#8276) Remove MaRuKu dependency
MaRuKu is GPLv2.  We're using it in a very small capacity
as a renderer for a CHANGELOG file that we don't really
need to show.  This patch removes the use of MaRuKu and
updates the application to no longer bundle MaRuKu.

The Version string on the normal navigation page
now will take you to the github project link.

Remove controller actions to ensure there isn't a helper
or anything related to displaying the changelog or
MaRuKu any more.

Reviewed-by: Jacob Helwig <jacob@puppetlabs.com>

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>