(#8316) Ruby sorting for ResourceEvents.
While most of our sorting should go on in the database, there is a complex
"natural" order for ResourceEvents: the 'ensure' property comes first, then
everything else should be alphabetical.

Since the set of events for a resource_status is usually handled distinctly,
and is always small in that context, it can make sense to handle that at the
Ruby level.

It is also quite complex to push that logic down into the database, since SQL
does not easily express the logic required.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>