(#6840) Remove need for VERSION file in puppet-dashboard
Instead of always relying on the VERSION file being present, we now
first test to see if it is, then fall back to using the output of
git-describe to determine which version of the Dashboard we are running.

Also, instead of just linking to the puppet-dashboard project on GitHub,
we now link to the CHANGELOG file for the version we are running.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>
Signed-off-by: Jacob Helwig <jacob@puppetlabs.com>