(#7567) Refactor dashboard packaging to allow for nightly builds
We now allow for builds out of master that utilizes git-describe for
versioning.  The generation of the tarball or package will create a
VERSION file with the proper contents.

The package/tarball will also use erb to substitute versions, dates and
other metadata into the Red Hat spec or Debian changelog.

The Debian package also changed to model more the builds from PE. This
means a newer rules configuration and the addition of the rmshebang
script to comply with Debian packaging guidelines.

The package:deb, package:rpm, and package:srpm rake tasks now use the
tarball generated from the package:tar task, which means that the spec,
changelog, and other packaging inputs are generated from their erb
templates.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>