(#8694) Add backtrace info to DelayedJobFailure
If reports fail to import with DelayedJob a failure entry is created
that has the error message, but did not have the backtrace, which often
is much more helpful for debugging.

This commit adds the backtrace to the UI in an expandable form so that
it doesn't look to spammy on the web page.

Reviewed-by: Pieter van de Bruggen <pieter@puppetlabs.com>