Add document outlining preferred contribution methods
We have historically had the preferred contribution process on the
Redmine wiki, however this is not obvious to people that don't already
know it is there.  By adding this document to the repository itself,
it becomes much easier for new contributors to find what the preferred
contribution methods are.

By having the preferred contribution method in the repository also
means that it becomes a "curated" document, which must go through the
same submission/review process that other changes to the repositories
go through.

Reviewed-by: Nick Fagerlund <nick.fagerlund@puppetlabs.com>
Reviewed-by: Nick Lewis <nick@puppetlabs.com>
Signed-off-by: Jacob Helwig <jacob@puppetlabs.com>