Allow setting of RUBY for the workers on redhat systems
While we are able to set an alternative RUBY path for the dashboard
server this is not possible for the workers.
We now include first the general dashboard sysconfig and afterwards
a (possible) dedicated workers sysconfig file, which could overwrite
the general dashboard settings.
This means that if you would adjust RUBY for the dashboard-server it
would also be used for the workers.