(#8748) Upgrade vendored daemons gem to fix umask on pids
We had version 1.1.0 of the daemons gem because that's what the older
version of delayed_job says it depends on.  However, since that time,
newer versions of daemons have been release with bug fixes, including
setting better permissions on pids and logs and not changing the umask

From vendor/gems/daemons-1.1.4/Releases

== Release 1.1.4: June 17, 2011

* Do not change the umask to 0000 when daemonizing anymore, just leave
* it as it.
  was (thanks to Jon Botelho).

== Release 1.1.0: June 20, 2010

* Set log files mode to 0644 (mikehale).
* Set pid file permissions to 0644 (mikehale).

I tested processing a lot of reports with delayed_job and the newer
daemons gem, started and stopped a bunch of delayed job workers, and saw
no problems, so feel comfortable vendoring this newer version of the
gem.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>