(#8785) - Revert "(#8748) Upgrade vendored daemons gem to fix umask on pids"
This reverts commit 651511cbafcbc9bd7794fdb6ddbc1802d01d35b5 and
3f88c7fb3231c09eabc67a79448d3a9c0eb03f2c

It turned out that the newer version of the daemons gem only worked
coincidentally with delayed_job when we had a plugin installed that used
the Registry in it's init file.  This caused a Dir to be opened in
lib/registry.rb:54.  Somehow, this prevented daemons from closing the
MySQL socket file, and so delayed_jobs worked.  The entirety of the
interactions are not understood, but we get development.log messages
like:

    #<ActiveRecord::StatementInvalid: Mysql::Error: MySQL server has gone
    away: UPDATE `delayed_jobs` SET locked_by = null, locked_at = nu
    ll WHERE (locked_by = 'delayed_job host:mattmac.local pid:72614')

and the offending lines in the deamons gem were narrowed down to
vendor/gems/daemons-1.1.4/lib/daemons/daemonize.rb:191.