(#8748) Put sensible umask on pids and logs that delayed_job creates
While it's preferable to NOT modify vendored code, having world
writeable pids and logs isn't ideal either.  This has been fixed in a
newer version of the daemons gem, but there are other issues getting
that to work, so for now we'll make a very minor change to set a more
sensible umask.