(#8790) Fix reports page column display and alignment
The columns on the reports page didn't line up because we weren't
showing pending report counts.  Also, the counts were showing as decimal
numbers when they are really all integer results.  Unfortunately, the
metrics table those numbers come from is generalized so that you can
store any kind of number, so the results default to displaying as
decimal, and the only real way to fix that display is in the view.