(#8796) Re-write misleading 500 error message
Previously, the 500 error message read, "We've been notified about this issue
and we'll take a look at it shortly." In fact, Dashboard was not notifying
anyone, and "we" certainly weren't coming to look at it. This commit replaces
the error message with practical (albeit basic) first steps for both technical
and non-technical users.