(#7934) Only link old file checksum to file bucket when contents have changed
The file bucket is only populated with the old contents of a file,
when the file is changed by Puppet.  Since the new contents of the
file are not put into the file bucket, the links that were shown for
the new file contents would only work if there was a file previously
had that particular checksum.

Now, we only make the checksum a link to view the contents from the
file bucket if the checksum is for the old file contents, and do not
show the link for the new file contents, or if the file was created in
that puppet run.

This does mean that we are hiding the link in cases where there is a
chance that the file contents could actually be shown.  For example,
if we are looking at an old report where a file was created, but the
contents were changed in later Puppet runs.