(#7934) Don't link md5s for new content
When running puppet agent when remote filebucketing is enabled, the old
content is file bucketed when there's a change, but not the new content.
Therefore, we won't link to the new content.

There's still plenty of potential to be wrong with the link.  The new
content would be there if an inspect run was done with --archive_files
after the apply report, and then we wouldn't link even though the
content was there.

We also will link when remote filebucketing isn't being used by the
agents (which may be more often than we thought since there's a bug
preventing it from working with the agent (#5362), or if the inspect
runs don't have --archive_files set (which defaults to false).

The baseline plugin will update its usage of popup_md5s as part of
ticket #9105.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>